More (wrap-text) changes
Got rid of (join-text) for now; as an input to (wrap-lines), (str/join)
works just as well.

Restored (wrap-text)'s old multi-variadic API, but instead of one
string you must pass a vector of strings.  That's not quite right,
either.  It should be a vector of strings, or a string.