Room description changes
Added a new room description hook, the
fact-based-room-description-hook.  It allows the description to have
parts that are only displayed if particular facts are true.

Made `define-room` return the room keyword; this is more pleasant at
the REPL than the entire @rooms map.

Updated the existing rooms with conditional parts to use the new hook.